gateway-api/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.5
httpx==0.27.0
python-dotenv==1.0.1
PyYAML==6.0.2
azure-identity==1.16.1
azure-keyvault-secrets==4.8.0
